refactor: publication on listings
