+++ monitor/reducers/publicationReducer.js
@@ -1,7 +1,7 @@
 import { eth, txUtils } from 'decentraland-eth'
 import { Log } from 'decentraland-commons'
 
-import { Publication, PublicationService } from '../../src/Publication'
+import { Publication, PublicationService } from '../../src/Listing'
 import { BlockTimestampService } from '../../src/BlockTimestamp'
 import { Tile } from '../../src/Tile'
 import { contractAddresses, eventNames } from '../../src/ethereum'
