+++ sanity/doctors/PublicationDoctor.js
@@ -4,7 +4,7 @@ import { Doctor } from './Doctor'
 import { Diagnosis } from './Diagnosis'
 import { asyncBatch } from '../../src/lib'
 import { Parcel, Estate } from '../../src/Asset'
-import { Publication } from '../../src/Publication'
+import { Publication } from '../../src/Listing'
 import { BlockchainEvent } from '../../src/BlockchainEvent'
 import { isParcel } from '../../shared/parcel'
 import { LISTING_STATUS } from '../../shared/listing'
