+++ scripts/mktcli.js
@@ -6,7 +6,7 @@ import { Log, cli } from 'decentraland-commons'
 import { SQL, db } from '../src/database'
 import { connectEth } from '../src/ethereum'
 import { Parcel, Estate } from '../src/Asset'
-import { Publication } from '../src/Publication'
+import { Publication } from '../src/Listing'
 import { BlockchainEvent } from '../src/BlockchainEvent'
 import { eventNames } from '../src/ethereum'
 import { mockModelDbOperations } from '../specs/utils'
