+++ src/Listing/Publication/Publication.spec.js
@@ -2,8 +2,8 @@ import { expect } from 'chai'
 
 import { Publication } from './Publication.model'
 import { PublicationService } from './Publication.service'
-import { Parcel, Estate } from '../Asset'
-import { ASSET_TYPES } from '../shared/asset'
+import { Parcel, Estate } from '../../Asset'
+import { ASSET_TYPES } from '../../shared/asset'
 
 describe('Publication', function() {
   describe('.primaryKey', function() {
