+++ src/Listing/index.js
@@ -1,3 +1,4 @@
 export { Listing } from './Listing'
 
 export * from './Bid'
+export * from './Publication'
