+++ src/ReqQueryParams/AssetReqQueryParams.js
@@ -1,6 +1,6 @@
 import { ReqQueryParams } from './ReqQueryParams'
 import { MarketplaceReqQueryParams } from './MarketplaceReqQueryParams'
-import { Publication } from '../Publication'
+import { Listing } from '../Listing'
 import { LISTING_ASSET_TYPES } from '../shared/listing'
 
 export class AssetReqQueryParams {
@@ -15,7 +15,7 @@ export class AssetReqQueryParams {
       ...queryParams
     } = this.marketplaceReqQueryParams.sanitize()
 
-    queryParams.asset_type = Publication.isValidAssetType(asset_type)
+    queryParams.asset_type = Listing.isValidAssetType(asset_type)
       ? asset_type
       : LISTING_ASSET_TYPES.parcel
