+++ src/Tile/Tile.model.js
@@ -4,7 +4,7 @@ import { TileAttributes } from './TileAttributes'
 import { Asset, Parcel, Estate } from '../Asset'
 import { Contribution } from '../Contribution'
 import { District } from '../District'
-import { Publication } from '../Publication'
+import { Publication } from '../Listing'
 import { SQL, raw } from '../database'
 import { asyncBatch } from '../lib'
 import { isDistrict } from '../../shared/district'
